requirements.txt
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.1
Pillow==5.4.1
pytz==2018.9
